harness_generator/src/requirements.txt
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
python-dotenv==1.2.1
PyYAML==6.0.3
smmap==5.0.2
typing_extensions==4.15.0
